requirements.txt
autopep8==1.6.0
black==21.12b0
click==8.0.3
Django==2.2.25
django-cors-headers==3.10.1
django-filter==21.1
djangorestframework==3.13.1
Faker==12.0.0
importlib-metadata==4.10.0
isort==5.10.1
Markdown==3.3.6
mixer==7.2.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
pycodestyle==2.8.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.3
typing_extensions==4.0.1
zipp==3.6.0
